Bird species dataset

Prepares the bird species dataset available in Kaggle here

bird_species_dataset(
  root,
  token = NULL,
  split = "train",
  download = FALSE,
  ...
)

Arguments

root
path to the data location

token
a path to the json file obtained in Kaggle. See here for additional info.

split
train, test or valid

download
wether to download or not

...
other arguments passed to torchvision::image_folder_dataset().

Value

A torch::dataset() ready to be used with dataloaders.

Details

We use pins for downloading and managing authetication. If you want to download the dataset you need to register the Kaggle board as described in this link. or pass the token argument.

Examples

if (torch::torch_is_installed() && FALSE) {
birds <- bird_species_dataset("./data", token = "path/to/kaggle.json",
                              download = TRUE)
length(birds)
}