Cityscapes Pix2Pix dataset

Downloads and prepares the cityscapes dataset that has been used in the pix2pix paper.

cityscapes_pix2pix_dataset(
  root,
  split = "train",
  download = FALSE,
  ...,
  transform = NULL,
  target_transform = NULL
)

Arguments

root

path to the data location

split

train, test or valid

download

wether to download or not

...

Currently unused.

transform

A function/transform that takes in an PIL image and returns a transformed version. E.g, transform_random_crop().

target_transform

A function/transform that takes in the target and transforms it.

Details

Find more information in the project website